Fix a subtle bug in how memory is managed between batches of curl body.
When a string or some other heap-allocated value is parsed early in a record, then the record is incomplete at the end of the batch of records, the memory context holding the value is reset. This causes a segfault when it is accessed to insert into the tuple. As a solution, we call datumCopy on all these values in a long-term memory context at the time of batch switching. This allows a few values to be allocated into long-term memory and not get destroyed when the batch memory is reset.